Move extract_level_gid, but also make it operate on id strings, not dicts
