+++ tests_v2/unit/app/utils/test_geostore.py
@@ -9,6 +9,7 @@ from app.errors import InvalidResponseError, RecordNotFoundError
 from app.models.enum.geostore import GeostoreOrigin
 from app.models.pydantic.geostore import GeostoreCommon
 from app.utils import geostore
+from app.utils.gadm import extract_level_gid
 from tests_v2.fixtures.sample_rw_geostore_response import geostore_common
 
 
@@ -90,3 +91,24 @@ async def test_get_geostore_mixed_errors(monkeypatch: MonkeyPatch):
             geostore_id_uuid, geostore_origin=GeostoreOrigin.rw
         )
     assert e.value.status_code == 500
+
+
+@pytest.mark.asyncio
+async def test_extract_level_gid() -> None:
+    # Normal gid values
+    match1 = "USA.5.10_1"
+    assert extract_level_gid(0, match1) == "USA"
+    assert extract_level_gid(1, match1) == "5"
+    assert extract_level_gid(2, match1) == "10"
+
+    # Ghana values with bad formatting (missing dot after GHA in gadm 4.1)
+    match2 = "GHA7.1_2"
+    assert extract_level_gid(0, match2) == "GHA"
+    assert extract_level_gid(1, match2) == "7"
+    assert extract_level_gid(2, match2) == "1"
+
+    # Indonesia values with bad formatting (missing suffix _1 in gadm 4.1)
+    match3 = "IDN.35.4"
+    assert extract_level_gid(0, match3) == "IDN"
+    assert extract_level_gid(1, match3) == "35"
+    assert extract_level_gid(2, match3) == "4"
